Turn LEDs off after longpress
Allow to turn off LEDs at any state immediately by long-pressing
button for a configurable amount of time.

Also make wipe-delay while switching configurable, so it's easier
to tune for LED rings with more LEDs (I'm using a 16 LED ring and
it would take 1280ms with the previous code code when switching, so
rather specify the wipe time to be constant and changeable).